% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{as_sparse_numeric}
\alias{as_sparse_numeric}
\title{Convert Numeric Vector to Sparse Numeric}
\arguments{
\item{from}{A \code{sparse_numeric} object.}
}
\value{
A \code{sparse_numeric} object.

A full numeric vector.
}
\description{
Converts a numeric vector into a \code{sparse_numeric} representation.

Converts a \code{sparse_numeric} object back into the full numeric vector.
}
