% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Compute mean of a sparse numeric vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{a sparse_numeric object}

\item{...}{Additional arguments (ignored)}
}
\value{
numeric mean
}
\description{
Compute mean of a sparse numeric vector
}
