% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm,sparse_numeric-method}
\alias{norm,sparse_numeric-method}
\title{Compute Euclidean norm of a sparse numeric vector}
\usage{
\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{a sparse_numeric object}

\item{...}{ignored}
}
\value{
numeric norm
}
\description{
Compute Euclidean norm of a sparse numeric vector
}
