% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm_sparse_numeric}
\alias{norm_sparse_numeric}
\title{L2 Norm of a Sparse Vector}
\usage{
\S4method{norm}{sparse_numeric,ANY}(x, type = "2", ...)
}
\arguments{
\item{x}{A \code{sparse_numeric}.}

\item{type}{Norm type. Only `"2"` supported.}
}
\value{
Numeric scalar.
}
\description{
L2 Norm of a Sparse Vector
}
