% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot Sparse Numeric Vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object.}
}
\description{
Plots two sparse numeric vectors on the same coordinate axes.
}
