% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_norm}
\alias{sparse_norm}
\alias{sparse_norm,sparse_numeric-method}
\title{L2 norm for sparse_numeric vectors (sparse_norm)}
\usage{
sparse_norm(x)

\S4method{sparse_norm}{sparse_numeric}(x)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}
}
\value{
Numeric scalar, the Euclidean norm of \code{x}.
}
\description{
L2 norm for sparse_numeric vectors (sparse_norm)
}
