% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_norm2}
\alias{sparse_norm2}
\title{Squared L2 Norm of a Sparse Vector}
\usage{
sparse_norm2(x)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}
}
\value{
Numeric scalar.
}
\description{
Squared L2 Norm of a Sparse Vector
}
