% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse numeric vector class}
\description{
An S4 class to represent numeric vectors in sparse form.
}
\section{Slots}{

\describe{
\item{\code{value}}{Numeric vector of non-zero values.}

\item{\code{pos}}{Integer vector of positions (1-based) of non-zero values.}

\item{\code{length}}{Integer giving the total length of the vector (including zeros).}
}}

