% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_numeric-ops}
\alias{sparse_numeric-ops}
\alias{+,sparse_numeric,sparse_numeric-method}
\alias{\S4method{+}{sparse_numeric,sparse_numeric}}
\alias{-,sparse_numeric,sparse_numeric-method}
\alias{\S4method{-}{sparse_numeric,sparse_numeric}}
\alias{*,sparse_numeric,sparse_numeric-method}
\alias{\S4method{*}{sparse_numeric,sparse_numeric}}
\title{Arithmetic operators for sparse_numeric vectors}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1, e2}{Objects of class \code{sparse_numeric}.}
}
\description{
These methods allow you to use the standard arithmetic operators
\code{+}, \code{-}, and \code{*} with \code{sparse_numeric} vectors.
They dispatch to \code{sparse_add()}, \code{sparse_sub()}, and
\code{sparse_mult()} respectively.
}
