% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_numeric_ops}
\alias{sparse_numeric_ops}
\alias{sparse_add}
\alias{sparse_sub}
\alias{sparse_mult}
\alias{sparse_crossprod}
\alias{density_ratio}
\alias{+}
\alias{-}
\alias{*}
\alias{plot}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\alias{density_ratio,sparse_numeric-method}
\title{Arithmetic and Utility Methods for sparse_numeric}
\usage{
sparse_add(x, y, ...)

sparse_sub(x, y, ...)

sparse_mult(x, y, ...)

sparse_crossprod(x, y, ...)

density_ratio(x, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{density_ratio}{sparse_numeric}(x, ...)
}
\arguments{
\item{x, y}{sparse_numeric objects}

\item{...}{Additional arguments passed to methods.}
}
\description{
Documentation for all arithmetic operators, crossproducts,
density ratio, and plot methods for sparse_numeric.
}
\keyword{internal}
