% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_sub}
\alias{sparse_sub}
\title{Subtract Sparse Numeric Vectors}
\usage{
sparse_sub(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object.}
}
\value{
A \code{sparse_numeric} representing \code{x - y}.
}
\description{
Subtract Sparse Numeric Vectors
}
