% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize,sparse_numeric-method}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse numeric vector}
\usage{
\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{a sparse_numeric object}

\item{...}{ignored}
}
\value{
sparse_numeric standardized object
}
\description{
Standardize a sparse numeric vector
}
