% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\title{Standardize a Sparse Numeric Vector}
\usage{
standardize(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric}.}
}
\value{
A standardized \code{sparse_numeric}.
}
\description{
Centers and scales a sparse numeric vector without constructing the dense
version. Zeros are accounted for analytically.
}
