% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_add}
\alias{sparse_add}
\title{Add Two Sparse Numeric Vectors}
\usage{
sparse_add(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object.}
}
\value{
A \code{sparse_numeric} representing \code{x + y}.
}
\description{
Add Two Sparse Numeric Vectors
}
