% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_add}
\alias{sparse_add}
\title{Add Two Sparse Numeric Vectors}
\usage{
sparse_add(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric vector.}

\item{y}{A sparse_numeric vector.}
}
\value{
A sparse_numeric representing the sum.
}
\description{
Element-wise addition of two sparse vectors of equal length.
}
