% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\title{Sparse Crossproduct (Dot Product)}
\usage{
sparse_crossprod(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric}.}

\item{y}{A \code{sparse_numeric}.}
}
\value{
A numeric scalar.
}
\description{
Sparse Crossproduct (Dot Product)
}
