% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_mult}
\alias{sparse_mult}
\title{Elementwise Multiply Sparse Numeric Vectors}
\usage{
sparse_mult(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric}.}

\item{y}{A \code{sparse_numeric}.}
}
\value{
A \code{sparse_numeric} representing \code{x * y}.
}
\description{
Elementwise Multiply Sparse Numeric Vectors
}
