% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_mult}
\alias{sparse_mult}
\title{Multiply Two Sparse Numeric Vectors}
\usage{
sparse_mult(x, y, ...)
}
\description{
Computes element-wise products only where both vectors have nonzero entries.
}
