% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse_numeric vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}

\item{...}{Ignored.}
}
\value{
A standardized \code{sparse_numeric} vector.
}
\description{
Standardizes each element of a \code{sparse_numeric} vector by subtracting
the mean and dividing by the standard deviation (as in \code{sd()}).
}
